requirements.txt
fastapi
uvicorn
python-multipart
fpdf
python-docx
